Updated NuGet dependencies, fixed runtime errors caused by that

+++ src/NadekoBot/Modules/Permissions/Services/PermissionsService.cs
@@ -32,7 +32,7 @@ namespace Mitternacht.Modules.Permissions.Services
 
             using (var uow = _db.UnitOfWork)
             {
-                foreach (var x in uow.GuildConfigs.Permissionsv2ForAll(client.Guilds.ToArray().Select(x => (long)x.Id).ToList()))
+                foreach (var x in uow.GuildConfigs.Permissionsv2ForAll(client.Guilds.ToArray().Select(x => x.Id).ToList()))
                 {
                     Cache.TryAdd(x.GuildId, new PermissionCache {
                         Verbose = x.VerbosePermissions,
